Refactor from double to decimal for pricing

+++ PizzaBox.Domain/Interfaces/ISellable.cs
@@ -2,7 +2,7 @@ namespace PizzaBox.Domain.Interfaces
 {
     public interface ISellable
     {
-        double Price { get; set; }
+        decimal Price { get; set; }
 
         string Name { get; set; }
     }
